requirements.txt
flask
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
joblib
Werkzeug
flask-login
flask-sqlalchemy
fpdf2
google-generativeai
requests
gunicorn
